# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GrowthForge AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/analyze:
    post:
      operationId: analyzeWebsite
      tags: [analysis]
      summary: Trigger website analysis for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessAnalysis"
  /projects/{id}/analysis:
    get:
      operationId: getBusinessAnalysis
      tags: [analysis]
      summary: Get stored business analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessAnalysis"
  /projects/{id}/personas:
    get:
      operationId: listPersonas
      tags: [analysis]
      summary: List customer personas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of personas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerPersona"
    post:
      operationId: generatePersonas
      tags: [analysis]
      summary: Generate customer personas for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated personas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerPersona"

  # ── Strategy ───────────────────────────────────────────────────────────────
  /projects/{id}/strategy:
    get:
      operationId: getMarketingStrategy
      tags: [strategy]
      summary: Get marketing strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marketing strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketingStrategy"
    post:
      operationId: generateMarketingStrategy
      tags: [strategy]
      summary: Generate marketing strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketingStrategy"

  # ── Competitors ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AnalysisInput:
      type: object
      required: [websiteUrl]
      properties:
        websiteUrl:
          type: string
        prompt:
          type: string
    BusinessAnalysis:
      type: object
      required: [id, projectId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        status:
          type: string
        businessSummary:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        products:
          type: ["string", "null"]
        services:
          type: ["string", "null"]
        uniqueValueProposition:
          type: ["string", "null"]
        targetCustomers:
          type: ["string", "null"]
        idealCustomerProfile:
          type: ["string", "null"]
        customerPainPoints:
          type: ["string", "null"]
        brandVoice:
          type: ["string", "null"]
        brandPositioning:
          type: ["string", "null"]
        customerBenefits:
          type: ["string", "null"]
        purchaseTriggers:
          type: ["string", "null"]
        marketOpportunities:
          type: ["string", "null"]
        growthOpportunities:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CustomerPersona:
      type: object
      required: [id, projectId, name, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        age:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        income:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        interests:
          type: ["string", "null"]
        motivations:
          type: ["string", "null"]
        objections:
          type: ["string", "null"]
        buyingTriggers:
          type: ["string", "null"]
        buyingJourney:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Strategy ─────────────────────────────────────────────────────────────
    MarketingStrategy:
      type: object
      required: [id, projectId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        status:
          type: string
        positioningStatement:
          type: ["string", "null"]
        messagingFramework:
          type: ["string", "null"]
        brandVoiceGuide:
          type: ["string", "null"]
        seoStrategy:
          type: ["string", "null"]
        campaignStrategy:
          type: ["string", "null"]
        leadGenerationStrategy:
          type: ["string", "null"]
        funnelRecommendations:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    # ── Competitors ──────────────────────────────────────────────────────────
...