# lib/api-zod/src/generated/types/customerPersona.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowthForge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerPersona {
  id: number;
  projectId: number;
  name: string;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  occupation?: string | null;
  /** @nullable */
  income?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  interests?: string | null;
  /** @nullable */
  motivations?: string | null;
  /** @nullable */
  objections?: string | null;
  /** @nullable */
  buyingTriggers?: string | null;
  /** @nullable */
  buyingJourney?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}